'Hero' Animation Added
