+++ lib/pages/home_detail_page.dart
@@ -1,0 +1,28 @@
+import 'dart:ffi';
+import 'package:flutter/material.dart';
+import 'package:flutter_catalog/models/catalog.dart';
+import 'package:flutter_catalog/widgets/themes.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class homeDetailPage extends StatelessWidget {
+  final Item catalog;
+
+  const homeDetailPage({Key? key, required this.catalog}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      backgroundColor: MyTheme.creamColor,
+      body: SafeArea(
+        child: Column(
+          children: [
+            Hero(
+                tag: Key(catalog.id.toString()),
+                child: Image.network(catalog.image)),
+          ],
+        ).p16(),
+      ),
+    );
+  }
+}
