+++ lib/pages/home_page.dart
@@ -5,6 +5,9 @@ import 'package:flutter_catalog/models/catalog.dart';
 import 'package:flutter_catalog/widgets/themes.dart';
 import 'package:velocity_x/velocity_x.dart';
 
+import 'home_widgets/catalog_header.dart';
+import 'home_widgets/catalog_list.dart';
+
 class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
 
@@ -44,9 +47,7 @@ class _HomePageState extends State<HomePage> {
               if (CatalogModel.items.isNotEmpty)
                 CatalogList().expand()
               else
-                const Center(
-                  child: CircularProgressIndicator(),
-                )
+                const CircularProgressIndicator().centered().py16().expand(),
             ],
           ),
         ),
@@ -55,89 +56,6 @@ class _HomePageState extends State<HomePage> {
   }
 }
 
-class CatalogHeader extends StatelessWidget {
-  const CatalogHeader({Key? key}) : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        "Catalog App".text.xl5.bold.color(MyTheme.darkBluishColor).make(),
-        "Trending Products".text.xl2.bold.make(),
-        const Divider(),
-      ],
-    );
-  }
-}
 
-class CatalogList extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return ListView.builder(
-      shrinkWrap: true,
-      itemCount: CatalogModel.items.length,
-      itemBuilder: (context, index) {
-        final catalog = CatalogModel.items[index];
-        return CatalogItem(catalog: catalog);
-      },
-    );
-  }
-}
 
-class CatalogItem extends StatelessWidget {
-  final Item catalog;
-
-  const CatalogItem({Key? key, required this.catalog}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return VxBox(
-        child: Row(
-      children: [
-        CatalogImage(image: catalog.image).p12(),
-        Expanded(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              catalog.name.text.lg.bold.color(MyTheme.darkBluishColor).make(),
-              catalog.desc.text.textStyle(context.captionStyle!).make(),
-              ButtonBar(
-                buttonPadding: EdgeInsets.zero,
-                alignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  "₹ ${catalog.price}".text.bold.make(),
-                  ElevatedButton(
-                    style: ButtonStyle(
-                        backgroundColor: MaterialStateProperty.all(
-                          MyTheme.darkBluishColor,
-                        ),
-                        shape:
-                            MaterialStateProperty.all(const StadiumBorder())),
-                    onPressed: () {},
-                    child: "Buy".text.medium.heightTight.make(),
-                  ),
-                ],
-              ).pOnly(right: 15.0),
-            ],
-          ),
-        )
-      ],
-    )).white.roundedLg.square(100).make().py12();
-  }
-}
-
-class CatalogImage extends StatelessWidget {
-  final String image;
-
-  const CatalogImage({Key? key, required this.image}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return Image.network(image)
-        .box
-        .color(MyTheme.creamColor)
-        .make()
-        .wh24(context);
-  }
-}
