+++ lib/pages/home_widgets/catalog_list.dart
@@ -1,0 +1,72 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_catalog/models/catalog.dart';
+import 'package:flutter_catalog/pages/home_detail_page.dart';
+import 'package:flutter_catalog/widgets/themes.dart';
+import 'package:velocity_x/velocity_x.dart';
+import 'catalog_image.dart';
+
+class CatalogList extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return ListView.builder(
+      shrinkWrap: true,
+      itemCount: CatalogModel.items.length,
+      itemBuilder: (context, index) {
+        final catalog = CatalogModel.items[index];
+        return InkWell(
+            onTap: () => Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => homeDetailPage(catalog: catalog),
+                  ),
+                ),
+            child: CatalogItem(catalog: catalog));
+      },
+    );
+  }
+}
+
+class CatalogItem extends StatelessWidget {
+  final Item catalog;
+
+  const CatalogItem({Key? key, required this.catalog}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return VxBox(
+        child: Row(
+      children: [
+        Hero(
+            tag: Key(catalog.id.toString()),
+            child: CatalogImage(image: catalog.image).p12()),
+        Expanded(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              catalog.name.text.lg.bold.color(MyTheme.darkBluishColor).make(),
+              catalog.desc.text.textStyle(context.captionStyle!).make(),
+              ButtonBar(
+                buttonPadding: EdgeInsets.zero,
+                alignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  "₹ ${catalog.price}".text.bold.make(),
+                  ElevatedButton(
+                    style: ButtonStyle(
+                        backgroundColor: MaterialStateProperty.all(
+                          MyTheme.darkBluishColor,
+                        ),
+                        shape:
+                            MaterialStateProperty.all(const StadiumBorder())),
+                    onPressed: () {},
+                    child: "Buy".text.medium.heightTight.make(),
+                  ),
+                ],
+              ).pOnly(right: 15.0),
+            ],
+          ),
+        )
+      ],
+    )).white.roundedLg.square(100).make().py12();
+  }
+}
