+++ lib/utils/routes.dart
@@ -1,5 +1,5 @@
 class MyRoutes {
   static String longinRoute = "/login";
   static String homeRoute = "/home";
-  
+  static String homeDetailsRoute = "/detail";
 }
